open a source member which contain period in its name
	new file:   .gitignore
	modified:   src/main/java/org/egomez/irpgeditor/DialogMemberNew.java
	new file:   src/main/java/org/egomez/irpgeditor/JobMemberCopy.java
	modified:   src/main/java/org/egomez/irpgeditor/Member.java
