+++ dependency-reduced-pom.xml
@@ -58,7 +58,7 @@
     </repository>
   </repositories>
   <properties>
-    <jdk.version>1.7</jdk.version>
+    <jdk.version>1.8</jdk.version>
   </properties>
 </project>
